+++ tests/gmprocess/io/asdf/path_utils_test.py
@@ -1,20 +1,14 @@
 from gmprocess.io.asdf import path_utils
-from gmprocess.io.read import read_data
-from gmprocess.utils.test_utils import read_data_dir
-from gmprocess.utils.config import get_config
 
 
-def test_path_utils():
+def test_path_utils(load_data_us1000778i, config):
     """Unit tests for path utils."""
     EVENTID = "us1000778i"
     LABEL = "ptest"
-    datafiles, event = read_data_dir("geonet", EVENTID, "*.V1A")
-    # select NZ.HSES
-    datafile = [df for df in datafiles if "HSES" in df][0]
-    stream = read_data(datafile)[0]
+    streams, _ = load_data_us1000778i
+    stream = streams.select(station="HSES")[0]
     trace = [tr for tr in stream if tr.stats.channel == "HN1"][0]
     tag = f"{EVENTID}_{LABEL}"
-    config = get_config()
 
     trace_path = path_utils.get_trace_path(trace, tag)
     trace_name = path_utils.get_trace_name(trace, tag)