+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## main
 
+- Adjust io tests to make use of pytest fixtures and get a bit of speedup.
+
 ## 1.3.0 / 2023-11-18
 
 - Add unprocessed waveform to summary plots.